requirements.txt
feedparser==6.0.11
groq==0.11.0
python-telegram-bot==21.6
aiohttp==3.10.5
